+++ test/t5_encoding_objects_test.dart
@@ -4,14 +4,15 @@ import 'package:test/test.dart';
 void main() {
   group('Object encode method', () {
     test(
-      'fail test: .encodeObject() with null imputs #1',
+      'success test: encode an object with null input #1',
       () {
         // prepare input
-        final data = null;
+        final User? data = null;
         // execute
-        final testCall = () => JsonString.encodeObject(data);
+        final jsonString = JsonString.encodeObject<User?>(data);
         // check
-        expect(testCall, throwsA(TypeMatcher<AssertionError>()));
+        expect(jsonString, isNotNull);
+        expect(jsonString, TypeMatcher<JsonString>());
       },
     );
     test(
@@ -40,7 +41,20 @@ void main() {
       },
     );
     test(
-      'fail test: try to encode a not-Jsonable object #1 (w/out encoder)',
+      'success test: encode a not Jsonable object #1 (w/out encoder + no check)',
+      () {
+        // prepare input
+        final data = Car(brand: 'Fiat', model: 'Punto');
+        // execute
+        final jsonString =
+            JsonString.encodeObject<Car>(data, checkIfJsonable: false);
+        // check
+        expect(jsonString, isNotNull);
+        expect(jsonString, TypeMatcher<JsonString>());
+      },
+    );
+    test(
+      'fail test: try to encode a not-Jsonable object #1 (w/out encoder + check)',
       () {
         // prepare input
         final data = Car(brand: 'Fiat', model: 'Punto');
@@ -56,8 +70,11 @@ void main() {
         // prepare input
         final data = Car(brand: 'Fiat', model: 'Punto');
         // execute
-        final testCall = () =>
-            JsonString.encodeObject<Car>(data, encoder: (car) => car.toJson());
+        final testCall = () => JsonString.encodeObject<Car>(data,
+            encoder: (car) => <String, dynamic>{
+                  'brand': car.brand,
+                  'model': car.model,
+                });
         final jsonString = testCall();
         // check
         expect(testCall, isNot(throwsA(TypeMatcher<JsonEncodingError>())));
@@ -69,7 +86,7 @@ void main() {
       'fail test: try to encode a primitive datatype #1',
       () {
         // prepare input
-        final data = "Hello!";
+        final data = 'Hello!';
         // execute
         final testCall = () => JsonString.encodeObject(data);
         // check
@@ -78,17 +95,6 @@ void main() {
     );
   });
   group('Object-list encode method', () {
-    test(
-      'fail test: .encodeObjectList() with null imputs #1',
-      () {
-        // prepare input
-        final data = null;
-        // execute
-        final testCall = () => JsonString.encodeObjectList(data);
-        // check
-        expect(testCall, throwsA(TypeMatcher<AssertionError>()));
-      },
-    );
     test(
       'success test: encode an object list #1 (Jsonable w/out encoder)',
       () {
@@ -111,8 +117,36 @@ void main() {
         final user2 = User(username: 'clara_brothers', password: 'qwerty');
         final data = [user1, user2];
         // execute
+        final jsonString = JsonString.encodeObjectList(data,
+            encoder: (User user) => user.toJson());
+        // check
+        expect(jsonString, isNotNull);
+        expect(jsonString, TypeMatcher<JsonString>());
+      },
+    );
+    test(
+      'fail test: try to encode a not-Jsonable object list #1 (w/out encoder + check)',
+      () {
+        // prepare input
+        final car1 = Car(brand: 'Fiat', model: 'Punto');
+        final car2 = Car(brand: 'Audi', model: 'A4');
+        final data = [car1, car2];
+        // execute
+        final testCall = () => JsonString.encodeObjectList(data);
+        // check
+        expect(testCall, throwsA(TypeMatcher<JsonEncodingError>()));
+      },
+    );
+    test(
+      'success test: try to encode a not-Jsonable object list #2  (w/out encoder + no check)',
+      () {
+        // prepare input
+        final car1 = Car(brand: 'Fiat', model: 'Punto');
+        final car2 = Car(brand: 'Audi', model: 'A4');
+        final data = [car1, car2];
+        // execute
         final jsonString =
-            JsonString.encodeObjectList(data, encoder: (user) => user.toJson());
+            JsonString.encodeObjectList(data, checkIfJsonable: false);
         // check
         expect(jsonString, isNotNull);
         expect(jsonString, TypeMatcher<JsonString>());
@@ -123,11 +157,12 @@ void main() {
       () {
         // prepare input
         final user1 = User(username: 'john_doe', password: 'querty');
-        final user2 = null;
-        final data = [user1, user2];
+        final data = <User?>[user1, null];
         // execute
-        final jsonString =
-            JsonString.encodeObjectList(data, encoder: (user) => user.toJson());
+        final jsonString = JsonString.encodeObjectList(
+          data,
+          encoder: (User? user) => user != null ? user.toJson() : null,
+        );
         // check
         expect(jsonString, isNotNull);
         expect(jsonString, TypeMatcher<JsonString>());
@@ -149,7 +184,7 @@ void main() {
       () {
         // prepare input
         final user1 = User(username: 'john_doe', password: 'querty');
-        final data = [0, user1, "2"];
+        final data = [0, user1, '2'];
         // execute
         final testCall = () => JsonString.encodeObjectList(data);
         // check
@@ -169,12 +204,12 @@ void main() {
       },
     );
     test(
-      'seccess test: try to encode a null filled list #2 (explicit type)',
+      'success test: try to encode a null filled list #2 (explicit type)',
       () {
         // prepare input
         final data = [null, null, null, null, null];
         // execute
-        final jsonString = JsonString.encodeObjectList<User>(data);
+        final jsonString = JsonString.encodeObjectList<User?>(data);
         // check
         expect(jsonString, isNotNull);
         expect(jsonString, TypeMatcher<JsonString>());
@@ -184,8 +219,8 @@ void main() {
 }
 
 class User with Jsonable {
-  String username;
-  String password;
+  String? username;
+  String? password;
 
   User({
     this.username,
@@ -194,40 +229,40 @@ class User with Jsonable {
 
   static User fromJson(Map<String, dynamic> json) {
     return User(
-      username: json['username'],
-      password: json['password'],
+      username: json['username'] as String?,
+      password: json['password'] as String?,
     );
   }
 
-  Map<String, dynamic> toJson() => {
+  @override
+  Map<String, dynamic> toJson() => <String, dynamic>{
         'username': username,
         'password': password,
       };
 
-  Map<String, dynamic> toSecureJson() => {
+  Map<String, dynamic> toSecureJson() => <String, dynamic>{
         'username': username,
-        'password': String.fromCharCodes(
-            List.filled(password.length, "*".codeUnitAt(0))),
+        'password': password != null
+            ? String.fromCharCodes(
+                List.filled(
+                  password!.length,
+                  '*'.codeUnitAt(0),
+                ),
+              )
+            : null,
       };
 }
 
 class Car {
-  String brand;
-  String model;
+  String? brand;
+  String? model;
 
   Car({
     this.brand,
     this.model,
   });
 
-  static Car fromJson(Map<String, dynamic> json) {
-    return Car(
-      brand: json['brand'],
-      model: json['model'],
-    );
-  }
-
-  Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => <String, dynamic>{
         'brand': brand,
         'model': model,
       };