+++ test/t8_decoding_objects_test.dart
@@ -2,21 +2,9 @@ import 'package:json_string/json_string.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('Object decode method', () {
+  group('Object (non-nullable) decode method', () {
     test(
-      'fail test: .decodeAsObject() with null input',
-      () {
-        // prepare input
-        final source = '{"username":"john_doe","password":"querty"}';
-        // execute
-        final jsonString = JsonString(source);
-        final testCall = () => jsonString.decodeAsObject(null);
-        // check
-        expect(testCall, throwsA(TypeMatcher<AssertionError>()));
-      },
-    );
-    test(
-      'success test: .decodeAsObject() with a proper decoder',
+      'success test: decode object with a proper decoder',
       () {
         // prepare input
         final source = '{"username":"john_doe","password":"querty"}';
@@ -27,6 +15,18 @@ void main() {
         // check
         expect(resultObject, isNotNull);
         expect(resultObject, TypeMatcher<User>());
+      },
+    );
+    test(
+      'fail test: this is null',
+      () {
+        // prepare input
+        final source = 'null';
+        // execute
+        final jsonString = JsonString(source);
+        final testCall = () => jsonString.decodeAsObject(User.fromJson);
+        // check
+        expect(testCall, throwsA(TypeMatcher<JsonDecodingError>()));
       },
     );
     test(
@@ -42,21 +42,54 @@ void main() {
       },
     );
   });
-  group('Object-list decode method', () {
+  group('Object (nullable) decode method', () {
     test(
-      'fail test: .decodeAsObjectList() with null input',
+      'success test: decode nullable object with a proper decoder',
       () {
         // prepare input
-        final source = '[{"username":"john_doe","password":"querty"}]';
+        final source = '{"username":"john_doe","password":"querty"}';
         // execute
         final jsonString = JsonString(source);
-        final testCall = () => jsonString.decodeAsObjectList(null);
+        final resultObject =
+            jsonString.decodeAsNullableObject(User.fromNullableJson);
+
         // check
-        expect(testCall, throwsA(TypeMatcher<AssertionError>()));
+        expect(resultObject, isNotNull);
+        expect(resultObject, TypeMatcher<User>());
       },
     );
     test(
-      'success test: .decodeAsObjectList() with a proper decoder',
+      'success test: this is null',
+      () {
+        // prepare input
+        final source = 'null';
+        // execute
+        final jsonString = JsonString(source);
+        final resultObject =
+            jsonString.decodeAsNullableObject(User.fromNullableJson);
+
+        // check
+        expect(resultObject, isNotNull);
+        expect(resultObject, TypeMatcher<User>());
+      },
+    );
+    test(
+      'fail test: no, this is not an object',
+      () {
+        // prepare input
+        final source = '["john_doe","querty"]';
+        // execute
+        final jsonString = JsonString(source);
+        final testCall =
+            () => jsonString.decodeAsNullableObject(User.fromNullableJson);
+        // check
+        expect(testCall, throwsA(TypeMatcher<JsonDecodingError>()));
+      },
+    );
+  });
+  group('Object-list (non-nullable) decode method', () {
+    test(
+      'success test: decode as object list with a proper decoder',
       () {
         // prepare input
         final source = '[{"username":"john_doe","password":"querty"}]';
@@ -71,6 +104,28 @@ void main() {
       },
     );
     test(
+      'fail test: this list contains null values',
+      () {
+        final source = '[{"username":"john_doe","password":"querty"}, null]';
+        // execute
+        final jsonString = JsonString(source);
+        final testCall = () => jsonString.decodeAsObjectList(User.fromJson);
+        // check
+        expect(testCall, throwsA(TypeMatcher<JsonDecodingError>()));
+      },
+    );
+    test(
+      'fail test: null-filled list',
+      () {
+        final source = '[null, null]';
+        // execute
+        final jsonString = JsonString(source);
+        final testCall = () => jsonString.decodeAsObjectList(User.fromJson);
+        // check
+        expect(testCall, throwsA(TypeMatcher<JsonDecodingError>()));
+      },
+    );
+    test(
       'fail test: try to decode a not-object list',
       () {
         final source = '[0, 1, 2, 3]';
@@ -81,26 +136,72 @@ void main() {
         expect(testCall, throwsA(TypeMatcher<JsonDecodingError>()));
       },
     );
+  });
+  group('Object-list (nullable) decode method', () {
     test(
-      'success test: .decodeAsObjectList() with a null filled list',
+      'success test: decode as nullable object list with a proper decoder',
       () {
         // prepare input
-        final source = '[null, null, null]';
+        final source = '[{"username":"john_doe","password":"querty"}]';
         // execute
         final jsonString = JsonString(source);
-        final objectList = jsonString.decodeAsObjectList(User.fromJson);
+        final resultObject =
+            jsonString.decodeAsNullableObjectList(User.fromNullableJson);
+
         // check
-        expect(objectList, isNotNull);
-        expect(objectList, hasLength(isNonZero));
-        expect(objectList, TypeMatcher<List<User>>());
+        expect(resultObject, isNotNull);
+        expect(resultObject, hasLength(isNonZero));
+        expect(resultObject, TypeMatcher<List<User>>());
+      },
+    );
+    test(
+      'success test: this list contains null values',
+      () {
+        final source = '[{"username":"john_doe","password":"querty"}, null]';
+        // execute
+        final jsonString = JsonString(source);
+        final resultObject =
+            jsonString.decodeAsNullableObjectList(User.fromNullableJson);
+
+        // check
+        expect(resultObject, isNotNull);
+        expect(resultObject, hasLength(isNonZero));
+        expect(resultObject, TypeMatcher<List<User>>());
+      },
+    );
+    test(
+      'success test: null-filled list',
+      () {
+        final source = '[null, null]';
+        // execute
+        final jsonString = JsonString(source);
+        final resultObject =
+            jsonString.decodeAsNullableObjectList(User.fromNullableJson);
+
+        // check
+        expect(resultObject, isNotNull);
+        expect(resultObject, hasLength(isNonZero));
+        expect(resultObject, TypeMatcher<List<User>>());
+      },
+    );
+    test(
+      'fail test: try to decode a not-object list',
+      () {
+        final source = '[0, 1, 2, 3]';
+        // execute
+        final jsonString = JsonString(source);
+        final testCall =
+            () => jsonString.decodeAsNullableObjectList(User.fromNullableJson);
+        // check
+        expect(testCall, throwsA(TypeMatcher<JsonDecodingError>()));
       },
     );
   });
 }
 
 class User {
-  String username;
-  String password;
+  String? username;
+  String? password;
 
   User({
     this.username,
@@ -109,16 +210,15 @@ class User {
 
   static User fromJson(Map<String, dynamic> json) {
     return User(
-      username: json['username'],
-      password: json['password'],
+      username: json['username'] as String?,
+      password: json['password'] as String?,
     );
   }
 
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is User &&
-          runtimeType == other.runtimeType &&
-          username == other.username &&
-          password == other.password;
+  static User fromNullableJson(Map<String, dynamic>? json) {
+    return User(
+      username: json?['username'] as String?,
+      password: json?['password'] as String?,
+    );
+  }
 }